+++ app/Main.hs
@@ -15,13 +15,20 @@ main = do
   (options, ()) <-
     simpleOptions
       $(simpleVersion Paths_PiG.version)
-      "Header for command line arguments"
-      "Program description, also for command line arguments"
+      "PiG - interpreter for a simple imperative language"
+      "start the interpreter"
       ( Options
           <$> switch
             ( long "verbose"
                 <> short 'v'
-                <> help "Verbose output?"
+                <> help "verbose output"
+            )
+          <*> strOption
+            ( long "file"
+                <> short 'f'
+                <> metavar "FILE"
+                <> value ""
+                <> help "start interpreter with FILE loaded"
             )
       )
       empty
@@ -34,6 +41,7 @@ main = do
             { appLogFunc = lf,
               appProcessContext = pc,
               appOptions = options,
-              appSettings = settings
+              appSettings = settings,
+              appVersion = Paths_PiG.version
             }
      in runRIO app run